+++ toolium/test/utils/test_dataset_map_param_context.py
@@ -90,6 +90,79 @@ def test_a_context_param_storage_and_feature_storage():
     assert expected_st == result_st
 
 
+def test_a_context_param_storage_and_run_storage():
+    """
+    Verification of a mapped parameter as CONTEXT saved in storage and run storage
+    """
+    class Context(object):
+        pass
+    context = Context()
+    context.attribute = "attribute value"
+    context.storage = {"storage_key": "storage entry value"}
+    context.run_storage = {"storage_key": "run storage entry value"}
+    dataset.behave_context = context
+
+    result_st = map_param("[CONTEXT:storage_key]")
+    expected_st = "storage entry value"
+    assert expected_st == result_st
+
+
+def test_store_key_in_feature_storage():
+    """
+    Verification of method store_key_in_storage with a mapped parameter as FEATURE saved in feature storage
+    """
+    class Context(object):
+        pass
+    context = Context()
+    context.attribute = "attribute value"
+    context.storage = {"storage_key": "storage entry value"}
+    context.feature_storage = {}
+    dataset.store_key_in_storage(context, "[FEATURE:storage_key]", "feature storage entry value")
+    dataset.behave_context = context
+
+    result_st = map_param("[CONTEXT:storage_key]")
+    expected_st = "feature storage entry value"
+    assert expected_st == result_st
+
+
+def test_store_key_in_run_storage():
+    """
+    Verification of method store_key_in_storage with a mapped parameter as RUN saved in run storage
+    """
+    class Context(object):
+        pass
+    context = Context()
+    context.attribute = "attribute value"
+    context.storage = {"storage_key": "storage entry value"}
+    context.run_storage = {}
+    context.feature_storage = {}
+    dataset.store_key_in_storage(context, "[RUN:storage_key]", "run storage entry value")
+    dataset.behave_context = context
+
+    result_st = map_param("[CONTEXT:storage_key]")
+    expected_st = "run storage entry value"
+    assert expected_st == result_st
+
+
+def test_a_context_param_using_store_key_in_storage():
+    """
+    Verification of a mapped parameter as CONTEXT saved in storage and run storage
+    """
+    class Context(object):
+        pass
+    context = Context()
+    context.attribute = "attribute value"
+    context.feature_storage = {}
+    context.storage = {"storage_key": "previous storage entry value"}
+    dataset.store_key_in_storage(context, "[FEATURE:storage_key]", "feature storage entry value")
+    dataset.store_key_in_storage(context, "[storage_key]", "storage entry value")
+    dataset.behave_context = context
+
+    result_st = map_param("[CONTEXT:storage_key]")
+    expected_st = "storage entry value"
+    assert expected_st == result_st
+
+
 def test_a_context_param_without_storage_and_feature_storage():
     """
     Verification of a mapped parameter as CONTEXT when before_feature and before_scenario have not been executed, so
@@ -356,6 +429,64 @@ def test_a_context_param_list_correct_index():
     assert map_param("[CONTEXT:list.cmsScrollableActions.1.id]") == 'ask-for-qa'
 
 
+def test_a_context_param_list_correct_negative_index():
+    """
+    Verification of a list with a correct negative index (In bounds) as CONTEXT
+    """
+    class Context(object):
+        pass
+    context = Context()
+
+    context.list = {
+        'cmsScrollableActions': [
+            {
+                'id': 'ask-for-duplicate',
+                'text': 'QA duplica'
+            },
+            {
+                'id': 'ask-for-qa',
+                'text': 'QA no duplica'
+            },
+            {
+                'id': 'ask-for-negative',
+                'text': 'QA negative index'
+            }
+        ]
+    }
+    dataset.behave_context = context
+    assert map_param("[CONTEXT:list.cmsScrollableActions.-1.id]") == 'ask-for-negative'
+    assert map_param("[CONTEXT:list.cmsScrollableActions.-3.id]") == 'ask-for-duplicate'
+
+
+def test_a_context_param_list_incorrect_negative_index():
+    """
+    Verification of a list with a incorrect negative index (In bounds) as CONTEXT
+    """
+    class Context(object):
+        pass
+    context = Context()
+
+    context.list = {
+        'cmsScrollableActions': [
+            {
+                'id': 'ask-for-duplicate',
+                'text': 'QA duplica'
+            },
+            {
+                'id': 'ask-for-qa',
+                'text': 'QA no duplica'
+            },
+            {
+                'id': 'ask-for-negative',
+                'text': 'QA negative index'
+            }
+        ]
+    }
+    with pytest.raises(Exception) as excinfo:
+        map_param("[CONTEXT:list.cmsScrollableActions.-5.id]")
+    assert "the expression '-5' was not able to select an element in the list" == str(excinfo.value)
+
+
 def test_a_context_param_list_oob_index():
     """
     Verification of a list with an incorrect index (Out of bounds) as CONTEXT
@@ -438,7 +569,6 @@ def test_a_context_param_class_no_numeric_index():
     context.list = ExampleClass()
     dataset.behave_context = context
 
-    print(context)
     with pytest.raises(Exception) as excinfo:
         map_param("[CONTEXT:list.cmsScrollableActions.prueba.id]")
     assert "the expression 'prueba' was not able to select an element in the list" == str(excinfo.value)