+++ toolium/test/utils/test_path_utils.py
@@ -30,6 +30,7 @@ filename_tests = (
     ('successful login -- @1.1 john.doe', 'successful_login_1_1_john_doe'),
     ('successful login -- @1.2 Mark options: {Length=10 Mark=mark File=file_name.jpg}',
      'successful_login_1_2_Mark_options___Length_10_Mark_mark_File_file_name_jpg'),
+    ('successful login -- @1.3 acción', 'successful_login_1_3_accion'),
 )